Add pseudo main code and init stack

+++ push_swap/include/converter.h
@@ -4,6 +4,6 @@
 # include "dependencies.h"
 # include "predicate.h"
 
-int	ft_atol(char *str);
+long long	ft_atol(char *str);
 
 #endif
